MISC: Move todo's to their own file
