+++ server/db.py
@@ -10,7 +10,8 @@ from sqlalchemy.engine import URL
 from models import Base
 from models.plant import Plant, Genus, Type
 from models.system import System, Light
-from models.alert import PlantAlert, Todo
+from models.alert import PlantAlert, Alert
+from models.todo import Todo, Task
 from models.mix import Soil, Mix
 
 import json
