+++ server/routes/todo_routes.py
@@ -4,7 +4,7 @@ from logger import logger
 from datetime import datetime
 
 from routes import GenericCRUD, APIBuilder
-from models.alert import Todo, Task
+from models.todo import Todo, Task
 
 bp = Blueprint('todos', __name__)
